tests/requirements-reliability-test.txt
pytest>=7.0.0
requests>=2.25.1
websocket-client>=1.0.0